fix: avoid Redux usage and related errors in non-newsletter email editors (#1688)
