+++ src/editor/mjml/index.js
@@ -3,19 +3,52 @@
 /**
  * WordPress dependencies
  */
+import { __ } from '@wordpress/i18n';
 import apiFetch from '@wordpress/api-fetch';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useEffect } from '@wordpress/element';
-import { refreshEmailHtml, usePrevious } from '../../newsletter-editor/utils';
+import { usePrevious } from '../../newsletter-editor/utils';
 
 /**
  * Internal dependencies
  */
-import { fetchNewsletterData, fetchSyncErrors, updateNewsletterDataError, updateIsRefreshingHtml, useIsRefreshingHtml } from '../../newsletter-editor/store';
 import { getServiceProvider } from '../../service-providers';
+import {
+	fetchNewsletterData,
+	fetchSyncErrors,
+	updateIsRefreshingHtml,
+} from '../../newsletter-editor/store';
+
+/**
+ * External dependencies
+ */
+import mjml2html from 'mjml-browser';
+
+/**
+ * Refresh the email-compliant HTML for a post.
+ *
+ * @param {number} postId      The current post ID.
+ * @param {string} postTitle   The current post title.
+ * @param {string} postContent The current post content.
+ * @return {Promise<string>} The refreshed email HTML.
+ */
+export const refreshEmailHtml = async ( postId, postTitle, postContent ) => {
+	const mjml = await apiFetch( {
+		path: `/newspack-newsletters/v1/post-mjml`,
+		method: 'POST',
+		data: {
+			post_id: postId,
+			title: postTitle,
+			content: postContent,
+		},
+	} );
+
+	// Once received MJML markup, convert it to email-compliant HTML and save as post meta.
+	const { html } = mjml2html( mjml, { keepComments: false, minify: true } );
+	return html;
+};
 
 function MJML() {
-	const isRefreshingHTML = useIsRefreshingHtml();
 	const {
 		saveSucceeded,
 		isPublishing,
@@ -54,7 +87,7 @@ function MJML() {
 			isAutosaveLocked: isPostAutosavingLocked(),
 		};
 	} );
-
+	const { createNotice } = useDispatch( 'core/notices' );
 	const { lockPostAutosaving, lockPostSaving, unlockPostSaving, editPost } = useDispatch(
 		'core/editor'
 	);
@@ -79,42 +112,43 @@ function MJML() {
 			! isSaving &&
 			! isAutosaving &&
 			! isPublishing &&
-			! isRefreshingHTML &&
 			! isSent &&
 			saveSucceeded
 		) {
-			updateIsRefreshingHtml( true );
-			lockPostSaving( 'newspack-newsletters-refresh-html' );
-			refreshEmailHtml( postId, postTitle, postContent )
-				.then( refreshedHtml => {
-					updateMetaValue( newspack_email_editor_data.email_html_meta, refreshedHtml );
-					return apiFetch( {
-						data: { meta: { [ newspack_email_editor_data.email_html_meta ]: refreshedHtml } },
-						method: 'POST',
-						path: `/wp/v2/${ postType }/${ postId }`,
-					} );
-				} ).then( () => {
-					// Rehydrate ESP newsletter data after completing sync.
-					if ( isSupportedESP ) {
-						return fetchNewsletterData( postId );
-					}
-					return true;
-				} ).then ( () => {
-					// Check for sync errors after refreshing the HTML.
-					if ( isSupportedESP ) {
-						return fetchSyncErrors( postId );
-					}
-					return true;
-				} )
-				.catch( e => {
-					updateNewsletterDataError( e );
-				} )
-				.finally( () => {
-					unlockPostSaving( 'newspack-newsletters-refresh-html' );
-					updateIsRefreshingHtml( false );
-				} );
+			refreshHtml();
 		}
 	}, [ isSaving, isAutosaving ] );
+
+	const refreshHtml = async () => {
+		try {
+			lockPostSaving( 'newspack-newsletters-refresh-html' );
+			if ( isSupportedESP ) {
+				updateIsRefreshingHtml( true );
+			}
+			const refreshedHtml = await refreshEmailHtml( postId, postTitle, postContent );
+			updateMetaValue( newspack_email_editor_data.email_html_meta, refreshedHtml );
+
+			// Save the refreshed HTML to post meta.
+			await apiFetch( {
+				data: { meta: { [ newspack_email_editor_data.email_html_meta ]: refreshedHtml } },
+				method: 'POST',
+				path: `/wp/v2/${ postType }/${ postId }`,
+			} );
+
+			// Rehydrate ESP newsletter data after completing sync.
+			if ( isSupportedESP ) {
+				await fetchNewsletterData( postId );
+				await fetchSyncErrors( postId );
+				updateIsRefreshingHtml( false );
+			}
+			unlockPostSaving( 'newspack-newsletters-refresh-html' );
+		} catch ( e ) {
+			createNotice( 'error', e?.message || __( 'Error refreshing email HTML.', 'newspack-newsletters' ), {
+				id: 'newspack-newsletters-mjml-error',
+				isDismissible: true,
+			} );
+		}
+	}
 }
 
 export default MJML;
